sayomi

+++ Assets/Scripts/Damage.cs
@@ -4,11 +4,11 @@ using UnityEngine;
 
 public class Damage
 {
-    public int amount;
+    public float amount;
     public DamageType type;
     public Transform source;
 
-    public Damage(int a, DamageType t, Transform s)
+    public Damage(float a, DamageType t, Transform s)
     {
         amount = a;
         type = t;
